Extract implementation of read, write and flush methods to TIOEngineBaseCommon
